Move to_grams() & to_meters() from utilities.utils to utilities.conversion
