+++ netbox/dcim/models/racks.py
@@ -17,8 +17,9 @@ from dcim.svg import RackElevationSVG
 from netbox.choices import ColorChoices
 from netbox.models import OrganizationalModel, PrimaryModel
 from netbox.models.features import ContactsMixin, ImageAttachmentsMixin
+from utilities.conversion import to_grams
 from utilities.fields import ColorField, NaturalOrderingField
-from utilities.utils import array_to_string, drange, to_grams
+from utilities.utils import array_to_string, drange
 from .device_components import PowerPort
 from .devices import Device, Module
 from .mixins import WeightMixin
